+++ CHANGELOG.md
@@ -21,6 +21,7 @@ v0.15.19 (??.??.????)
  * new Wrapper Class Tx_Rnbase_Interface_Singleton for \TYPO3\CMS\Core\SingletonInterface or t3lib_Singleton which is compatible for all TYPO3 versions
  * new Wrapper function tx_rnbase_util_Misc::logDeprecatedFunction
  * new Wrapper function tx_rnbase_util_Logger::devLog
+ * new Wrapper function tx_rnbase_mod_Util::mapRecordTypeToSpriteIconName
 
 v0.15.18 (02.12.2015)
  * Bugfix: tx_rnbase_util_Dates::date_mysql2tstamp liefert nun immer einen Integer und nicht NULL im Fehlerfall